+++ pkg/sql/opt/norm/general_funcs.go
@@ -15,9 +15,7 @@
 package norm
 
 import (
-	"math"
 	"reflect"
-	"sort"
 
 	"github.com/cockroachdb/apd"
 	"github.com/cockroachdb/cockroach/pkg/sql/coltypes"
@@ -29,9 +27,6 @@ import (
 	"github.com/cockroachdb/cockroach/pkg/sql/pgwire/pgerror"
 	"github.com/cockroachdb/cockroach/pkg/sql/sem/tree"
 	"github.com/cockroachdb/cockroach/pkg/sql/sem/types"
-	"github.com/cockroachdb/cockroach/pkg/util"
-	"github.com/cockroachdb/cockroach/pkg/util/json"
-	"github.com/cockroachdb/cockroach/pkg/util/log"
 )
 
 // CustomFuncs contains all the custom match and replace functions used by
@@ -54,71 +49,8 @@ func (c *CustomFuncs) Succeeded(result opt.Expr) bool {
 
 // ----------------------------------------------------------------------
 //
-// ScalarList functions
-//   General custom match and replace functions used to test and construct
-//   scalar lists.
-//
-// ----------------------------------------------------------------------
-
-// NeedSortedUniqueList returns true if the given list is composed entirely of
-// constant values that are either not in sorted order or have duplicates. If
-// true, then ConstructSortedUniqueList needs to be called on the list to
-// normalize it.
-func (c *CustomFuncs) NeedSortedUniqueList(list memo.ScalarListExpr) bool {
-	if len(list) <= 1 {
-		return false
-	}
-	ls := listSorter{cf: c, list: list}
-	var needSortedUniqueList bool
-	for i, item := range list {
-		if !opt.IsConstValueOp(item) {
-			return false
-		}
-		if i != 0 && !ls.less(i-1, i) {
-			needSortedUniqueList = true
-		}
-	}
-	return needSortedUniqueList
-}
-
-// ConstructSortedUniqueList sorts the given list and removes duplicates, and
-// returns the resulting list. See the comment for listSorter.compare for
-// comparison rule details.
-func (c *CustomFuncs) ConstructSortedUniqueList(
-	list memo.ScalarListExpr,
-) (memo.ScalarListExpr, types.T) {
-	// Make a copy of the list, since it needs to stay immutable.
-	newList := make(memo.ScalarListExpr, len(list))
-	copy(newList, list)
-	ls := listSorter{cf: c, list: newList}
-
-	// Sort the list.
-	sort.Slice(ls.list, ls.less)
-
-	// Remove duplicates from the list.
-	n := 0
-	for i := range newList {
-		if i == 0 || ls.compare(i-1, i) < 0 {
-			newList[n] = newList[i]
-			n++
-		}
-	}
-	newList = newList[:n]
-
-	// Construct the type of the tuple.
-	typ := types.TTuple{Types: make([]types.T, n)}
-	for i := range newList {
-		typ.Types[i] = newList[i].DataType()
-	}
-
-	return newList, typ
-}
-
-// ----------------------------------------------------------------------
-//
 // Typing functions
-//   General custom match and replace functions used to test and construct
-//   expression data types.
+//   General functions used to test and construct expression data types.
 //
 // ----------------------------------------------------------------------
 
@@ -135,11 +67,6 @@ func (c *CustomFuncs) HasColType(scalar opt.ScalarExpr, dstTyp coltypes.T) bool 
 // IsString returns true if the given scalar expression is of type String.
 func (c *CustomFuncs) IsString(scalar opt.ScalarExpr) bool {
 	return scalar.DataType() == types.String
-}
-
-// ColTypeToDatumType maps the given column type to a datum type.
-func (c *CustomFuncs) ColTypeToDatumType(colTyp coltypes.T) types.T {
-	return coltypes.CastTargetToDatumType(colTyp)
 }
 
 // BoolType returns the boolean SQL type.
@@ -175,11 +102,24 @@ func (c *CustomFuncs) BinaryColType(op opt.Operator, left, right opt.ScalarExpr)
 	return colType
 }
 
+// IsAdditive returns true if the type of the expression supports addition and
+// subtraction in the natural way. This differs from "has a +/- Numeric
+// implementation" because JSON has an implementation for "- INT" which doesn't
+// obey x - 0 = x. Additive types include all numeric types as well as
+// timestamps and dates.
+func (c *CustomFuncs) IsAdditive(e opt.ScalarExpr) bool {
+	return types.IsAdditiveType(e.DataType())
+}
+
+// ColTypeToDatumType maps the given column type to a datum type.
+func (c *CustomFuncs) ColTypeToDatumType(colTyp coltypes.T) types.T {
+	return coltypes.CastTargetToDatumType(colTyp)
+}
+
 // ----------------------------------------------------------------------
 //
-// Property functions
-//   General custom match and replace functions used to test expression
-//   logical properties.
+// Column functions
+//   General custom match and replace functions related to columns.
 //
 // ----------------------------------------------------------------------
 
@@ -194,12 +134,6 @@ func (c *CustomFuncs) OutputCols2(left, right memo.RelExpr) opt.ColSet {
 	return left.Relational().OutputCols.Union(right.Relational().OutputCols)
 }
 
-// CandidateKey returns the candidate key columns from the given input
-// expression. If there is no candidate key, CandidateKey returns ok=false.
-func (c *CustomFuncs) CandidateKey(input memo.RelExpr) (key opt.ColSet, ok bool) {
-	return input.Relational().FuncDeps.StrictKey()
-}
-
 // IsColNotNull returns true if the given input column is never null.
 func (c *CustomFuncs) IsColNotNull(col opt.ColumnID, input memo.RelExpr) bool {
 	return input.Relational().NotNullCols.Contains(int(col))
@@ -212,80 +146,9 @@ func (c *CustomFuncs) IsColNotNull2(col opt.ColumnID, left, right memo.RelExpr) 
 		right.Relational().NotNullCols.Contains(int(col))
 }
 
-// OuterCols returns the set of outer columns associated with the given
-// expression, whether it be a relational or scalar operator.
-func (c *CustomFuncs) OuterCols(e opt.Expr) opt.ColSet {
-	return c.sharedProps(e).OuterCols
-}
-
-// HasOuterCols returns true if the input expression has at least one outer
-// column, or in other words, a reference to a variable that is not bound within
-// its own scope. For example:
-//
-//   SELECT * FROM a WHERE EXISTS(SELECT * FROM b WHERE b.x = a.x)
-//
-// The a.x variable in the EXISTS subquery references a column outside the scope
-// of the subquery. It is an "outer column" for the subquery (see the comment on
-// RelationalProps.OuterCols for more details).
-func (c *CustomFuncs) HasOuterCols(input opt.Expr) bool {
-	return !c.OuterCols(input).Empty()
-}
-
-// IsBoundBy returns true if all outer references in the source expression are
-// bound by the given columns. For example:
-//
-//   (InnerJoin
-//     (Scan a)
-//     (Scan b)
-//     [ ... $item:(FiltersItem (Eq (Variable a.x) (Const 1))) ... ]
-//   )
-//
-// The $item expression is fully bound by the output columns of the (Scan a)
-// expression because all of its outer references are satisfied by the columns
-// produced by the Scan.
-func (c *CustomFuncs) IsBoundBy(src opt.Expr, cols opt.ColSet) bool {
-	return c.OuterCols(src).SubsetOf(cols)
-}
-
-// IsCorrelated returns true if any variable in the source expression references
-// a column from the destination expression. For example:
-//   (InnerJoin
-//     (Scan a)
-//     (Scan b)
-//     [ ... (FiltersItem $item:(Eq (Variable a.x) (Const 1))) ... ]
-//   )
-//
-// The $item expression is correlated with the (Scan a) expression because it
-// references one of its columns. But the $item expression is not correlated
-// with the (Scan b) expression.
-func (c *CustomFuncs) IsCorrelated(src, dst memo.RelExpr) bool {
-	return src.Relational().OuterCols.Intersects(dst.Relational().OutputCols)
-}
-
 // HasNoCols returns true if the input expression has zero output columns.
 func (c *CustomFuncs) HasNoCols(input memo.RelExpr) bool {
 	return input.Relational().OutputCols.Empty()
-}
-
-// HasZeroRows returns true if the input expression never returns any rows.
-func (c *CustomFuncs) HasZeroRows(input memo.RelExpr) bool {
-	return input.Relational().Cardinality.IsZero()
-}
-
-// HasOneRow returns true if the input expression always returns exactly one
-// row.
-func (c *CustomFuncs) HasOneRow(input memo.RelExpr) bool {
-	return input.Relational().Cardinality.IsOne()
-}
-
-// HasZeroOrOneRow returns true if the input expression returns at most one row.
-func (c *CustomFuncs) HasZeroOrOneRow(input memo.RelExpr) bool {
-	return input.Relational().Cardinality.IsZeroOrOne()
-}
-
-// CanHaveZeroRows returns true if the input expression might return zero rows.
-func (c *CustomFuncs) CanHaveZeroRows(input memo.RelExpr) bool {
-	return input.Relational().Cardinality.CanBeZero()
 }
 
 // ColsAreEmpty returns true if the column set is empty.
@@ -335,6 +198,218 @@ func (c *CustomFuncs) DifferenceCols(left, right opt.ColSet) opt.ColSet {
 	return left.Difference(right)
 }
 
+// MakeEmptyColSet returns a column set with no columns in it.
+func (c *CustomFuncs) MakeEmptyColSet() opt.ColSet {
+	return opt.ColSet{}
+}
+
+// FirstCol returns the first column in the input expression.
+func (c *CustomFuncs) FirstCol(in memo.RelExpr) opt.ColumnID {
+	inCol, _ := c.OutputCols(in).Next(0)
+	return opt.ColumnID(inCol)
+}
+
+// ----------------------------------------------------------------------
+//
+// Outer column functions
+//   General custom functions related to outer column references.
+//
+// ----------------------------------------------------------------------
+
+// OuterCols returns the set of outer columns associated with the given
+// expression, whether it be a relational or scalar operator.
+func (c *CustomFuncs) OuterCols(e opt.Expr) opt.ColSet {
+	return c.sharedProps(e).OuterCols
+}
+
+// HasOuterCols returns true if the input expression has at least one outer
+// column, or in other words, a reference to a variable that is not bound within
+// its own scope. For example:
+//
+//   SELECT * FROM a WHERE EXISTS(SELECT * FROM b WHERE b.x = a.x)
+//
+// The a.x variable in the EXISTS subquery references a column outside the scope
+// of the subquery. It is an "outer column" for the subquery (see the comment on
+// RelationalProps.OuterCols for more details).
+func (c *CustomFuncs) HasOuterCols(input opt.Expr) bool {
+	return !c.OuterCols(input).Empty()
+}
+
+// IsCorrelated returns true if any variable in the source expression references
+// a column from the destination expression. For example:
+//   (InnerJoin
+//     (Scan a)
+//     (Scan b)
+//     [ ... (FiltersItem $item:(Eq (Variable a.x) (Const 1))) ... ]
+//   )
+//
+// The $item expression is correlated with the (Scan a) expression because it
+// references one of its columns. But the $item expression is not correlated
+// with the (Scan b) expression.
+func (c *CustomFuncs) IsCorrelated(src, dst memo.RelExpr) bool {
+	return src.Relational().OuterCols.Intersects(dst.Relational().OutputCols)
+}
+
+// IsBoundBy returns true if all outer references in the source expression are
+// bound by the given columns. For example:
+//
+//   (InnerJoin
+//     (Scan a)
+//     (Scan b)
+//     [ ... $item:(FiltersItem (Eq (Variable a.x) (Const 1))) ... ]
+//   )
+//
+// The $item expression is fully bound by the output columns of the (Scan a)
+// expression because all of its outer references are satisfied by the columns
+// produced by the Scan.
+func (c *CustomFuncs) IsBoundBy(src opt.Expr, cols opt.ColSet) bool {
+	return c.OuterCols(src).SubsetOf(cols)
+}
+
+// AreProjectionsCorrelated returns true if any element in the projections
+// references any of the given columns.
+func (c *CustomFuncs) AreProjectionsCorrelated(
+	projections memo.ProjectionsExpr, cols opt.ColSet,
+) bool {
+	for i := range projections {
+		if projections[i].ScalarProps(c.mem).OuterCols.Intersects(cols) {
+			return true
+		}
+	}
+	return false
+}
+
+// IsZipCorrelated returns true if any element in the zip references
+// any of the given columns.
+func (c *CustomFuncs) IsZipCorrelated(zip memo.ZipExpr, cols opt.ColSet) bool {
+	for i := range zip {
+		if zip[i].ScalarProps(c.mem).OuterCols.Intersects(cols) {
+			return true
+		}
+	}
+	return false
+}
+
+// FilterOuterCols returns the union of all outer columns from the given filter
+// conditions.
+func (c *CustomFuncs) FilterOuterCols(filters memo.FiltersExpr) opt.ColSet {
+	var colSet opt.ColSet
+	for i := range filters {
+		colSet.UnionWith(filters[i].ScalarProps(c.mem).OuterCols)
+	}
+	return colSet
+}
+
+// FiltersBoundBy returns true if all outer references in any of the filter
+// conditions are bound by the given columns. For example:
+//
+//   (InnerJoin
+//     (Scan a)
+//     (Scan b)
+//     $filters:[ (FiltersItem (Eq (Variable a.x) (Const 1))) ]
+//   )
+//
+// The $filters expression is fully bound by the output columns of the (Scan a)
+// expression because all of its outer references are satisfied by the columns
+// produced by the Scan.
+func (c *CustomFuncs) FiltersBoundBy(filters memo.FiltersExpr, cols opt.ColSet) bool {
+	for i := range filters {
+		if !filters[i].ScalarProps(c.mem).OuterCols.SubsetOf(cols) {
+			return false
+		}
+	}
+	return true
+}
+
+// ProjectionOuterCols returns the union of all outer columns from the given
+// projection expressions.
+func (c *CustomFuncs) ProjectionOuterCols(projections memo.ProjectionsExpr) opt.ColSet {
+	var colSet opt.ColSet
+	for i := range projections {
+		colSet.UnionWith(projections[i].ScalarProps(c.mem).OuterCols)
+	}
+	return colSet
+}
+
+// AggregationOuterCols returns the union of all outer columns from the given
+// aggregation expressions.
+func (c *CustomFuncs) AggregationOuterCols(aggregations memo.AggregationsExpr) opt.ColSet {
+	var colSet opt.ColSet
+	for i := range aggregations {
+		colSet.UnionWith(aggregations[i].ScalarProps(c.mem).OuterCols)
+	}
+	return colSet
+}
+
+// ZipOuterCols returns the union of all outer columns from the given
+// zip expressions.
+func (c *CustomFuncs) ZipOuterCols(zip memo.ZipExpr) opt.ColSet {
+	var colSet opt.ColSet
+	for i := range zip {
+		colSet.UnionWith(zip[i].ScalarProps(c.mem).OuterCols)
+	}
+	return colSet
+}
+
+// ----------------------------------------------------------------------
+//
+// Row functions
+//   General custom match and replace functions related to rows.
+//
+// ----------------------------------------------------------------------
+
+// HasZeroRows returns true if the input expression never returns any rows.
+func (c *CustomFuncs) HasZeroRows(input memo.RelExpr) bool {
+	return input.Relational().Cardinality.IsZero()
+}
+
+// HasOneRow returns true if the input expression always returns exactly one
+// row.
+func (c *CustomFuncs) HasOneRow(input memo.RelExpr) bool {
+	return input.Relational().Cardinality.IsOne()
+}
+
+// HasZeroOrOneRow returns true if the input expression returns at most one row.
+func (c *CustomFuncs) HasZeroOrOneRow(input memo.RelExpr) bool {
+	return input.Relational().Cardinality.IsZeroOrOne()
+}
+
+// CanHaveZeroRows returns true if the input expression might return zero rows.
+func (c *CustomFuncs) CanHaveZeroRows(input memo.RelExpr) bool {
+	return input.Relational().Cardinality.CanBeZero()
+}
+
+// ----------------------------------------------------------------------
+//
+// Key functions
+//   General custom match and replace functions related to keys.
+//
+// ----------------------------------------------------------------------
+
+// CandidateKey returns the candidate key columns from the given input
+// expression. If there is no candidate key, CandidateKey returns ok=false.
+func (c *CustomFuncs) CandidateKey(input memo.RelExpr) (key opt.ColSet, ok bool) {
+	return input.Relational().FuncDeps.StrictKey()
+}
+
+// GroupingColsAreKey returns true if the input expression's grouping columns
+// form a strict key for its output rows. A strict key means that any two rows
+// will have unique key column values. Nulls are treated as equal to one another
+// (i.e. no duplicate nulls allowed). Having a strict key means that the set of
+// key column values uniquely determine the values of all other columns in the
+// relation.
+func (c *CustomFuncs) GroupingColsAreKey(grouping *memo.GroupingPrivate, input memo.RelExpr) bool {
+	colSet := grouping.GroupingCols
+	return input.Relational().FuncDeps.ColsAreStrictKey(colSet)
+}
+
+// ----------------------------------------------------------------------
+//
+// Property functions
+//   General custom functions related to expression logical properties.
+//
+// ----------------------------------------------------------------------
+
 // sharedProps returns the shared logical properties for the given expression.
 // Only relational expressions and certain scalar list items (e.g. FiltersItem,
 // ProjectionsItem, AggregationsItem) have shared properties.
@@ -351,7 +426,7 @@ func (c *CustomFuncs) sharedProps(e opt.Expr) *props.Shared {
 // ----------------------------------------------------------------------
 //
 // Ordering functions
-//   General custom match and replace functions related to orderings.
+//   General functions related to orderings.
 //
 // ----------------------------------------------------------------------
 
@@ -390,20 +465,9 @@ func (c *CustomFuncs) EmptyOrdering() physical.OrderingChoice {
 // -----------------------------------------------------------------------
 //
 // Filter functions
-//   General custom match and replace functions used to test and construct
-//   filters in Select and Join rules.
+//   General functions used to test and construct filters.
 //
 // -----------------------------------------------------------------------
-
-// FilterOuterCols returns the union of all outer columns from the given filter
-// conditions.
-func (c *CustomFuncs) FilterOuterCols(filters memo.FiltersExpr) opt.ColSet {
-	var colSet opt.ColSet
-	for i := range filters {
-		colSet.UnionWith(filters[i].ScalarProps(c.mem).OuterCols)
-	}
-	return colSet
-}
 
 // FilterHasCorrelatedSubquery returns true if any of the filter conditions
 // contain a correlated subquery.
@@ -478,27 +542,6 @@ func (c *CustomFuncs) ReplaceFiltersItem(
 	panic(pgerror.NewAssertionErrorf("item to replace is not in the list: %v", search))
 }
 
-// FiltersBoundBy returns true if all outer references in any of the filter
-// conditions are bound by the given columns. For example:
-//
-//   (InnerJoin
-//     (Scan a)
-//     (Scan b)
-//     $filters:[ (FiltersItem (Eq (Variable a.x) (Const 1))) ]
-//   )
-//
-// The $filters expression is fully bound by the output columns of the (Scan a)
-// expression because all of its outer references are satisfied by the columns
-// produced by the Scan.
-func (c *CustomFuncs) FiltersBoundBy(filters memo.FiltersExpr, cols opt.ColSet) bool {
-	for i := range filters {
-		if !filters[i].ScalarProps(c.mem).OuterCols.SubsetOf(cols) {
-			return false
-		}
-	}
-	return true
-}
-
 // ExtractBoundConditions returns a new list containing only those expressions
 // from the given list that are fully bound by the given columns (i.e. all
 // outer references are to one of these columns). For example:
@@ -543,223 +586,12 @@ func (c *CustomFuncs) ExtractUnboundConditions(
 	return newFilters
 }
 
-// CanConsolidateFilters returns true if there are at least two different
-// filter conditions that contain the same variable, where the conditions
-// have tight constraints and contain a single variable. For example,
-// CanConsolidateFilters returns true with filters {x > 5, x < 10}, but false
-// with {x > 5, y < 10} and {x > 5, x = y}.
-func (c *CustomFuncs) CanConsolidateFilters(filters memo.FiltersExpr) bool {
-	var seen opt.ColSet
-	for i := range filters {
-		if col, ok := c.canConsolidateFilter(&filters[i]); ok {
-			if seen.Contains(col) {
-				return true
-			}
-			seen.Add(col)
-		}
-	}
-	return false
-}
-
-// canConsolidateFilter determines whether a filter condition can be
-// consolidated. Filters can be consolidated if they have tight constraints
-// and contain a single variable. Examples of such filters include x < 5 and
-// x IS NULL. If the filter can be consolidated, canConsolidateFilter returns
-// the column ID of the variable and ok=true. Otherwise, canConsolidateFilter
-// returns ok=false.
-func (c *CustomFuncs) canConsolidateFilter(filter *memo.FiltersItem) (col int, ok bool) {
-	if !filter.ScalarProps(c.mem).TightConstraints {
-		return 0, false
-	}
-
-	outerCols := c.OuterCols(filter)
-	if outerCols.Len() != 1 {
-		return 0, false
-	}
-
-	col, _ = outerCols.Next(0)
-	return col, true
-}
-
-// ConsolidateFilters consolidates filter conditions that contain the same
-// variable, where the conditions have tight constraints and contain a single
-// variable. The consolidated filters are combined with a tree of nested
-// And operations, and wrapped with a Range expression.
-//
-// See the ConsolidateSelectFilters rule for more details about why this is
-// necessary.
-func (c *CustomFuncs) ConsolidateFilters(filters memo.FiltersExpr) memo.FiltersExpr {
-	// First find the columns that have filter conditions that can be
-	// consolidated.
-	var seen, seenTwice opt.ColSet
-	for i := range filters {
-		if col, ok := c.canConsolidateFilter(&filters[i]); ok {
-			if seen.Contains(col) {
-				seenTwice.Add(col)
-			} else {
-				seen.Add(col)
-			}
-		}
-	}
-
-	newFilters := make(memo.FiltersExpr, seenTwice.Len(), len(filters)-seenTwice.Len())
-
-	// newFilters contains an empty item for each of the new Range expressions
-	// that will be created below. Fill in rangeMap to track which column
-	// corresponds to each item.
-	var rangeMap util.FastIntMap
-	i := 0
-	for col, ok := seenTwice.Next(0); ok; col, ok = seenTwice.Next(col + 1) {
-		rangeMap.Set(col, i)
-		i++
-	}
-
-	// Iterate through each existing filter condition, and either consolidate it
-	// into one of the new Range expressions or add it unchanged to the new
-	// filters.
-	for i := range filters {
-		if col, ok := c.canConsolidateFilter(&filters[i]); ok && seenTwice.Contains(col) {
-			// This is one of the filter conditions that can be consolidated into a
-			// Range.
-			cond := filters[i].Condition
-			switch t := cond.(type) {
-			case *memo.RangeExpr:
-				// If it is already a range expression, unwrap it.
-				cond = t.And
-			}
-			rangeIdx, _ := rangeMap.Get(col)
-			rangeItem := &newFilters[rangeIdx]
-			if rangeItem.Condition == nil {
-				// This is the first condition.
-				rangeItem.Condition = cond
-			} else {
-				// Build a left-deep tree of ANDs.
-				rangeItem.Condition = c.f.ConstructAnd(rangeItem.Condition, cond)
-			}
-		} else {
-			newFilters = append(newFilters, filters[i])
-		}
-	}
-
-	// Construct each of the new Range operators now that we have built the
-	// conjunctions.
-	for i, n := 0, seenTwice.Len(); i < n; i++ {
-		newFilters[i].Condition = c.f.ConstructRange(newFilters[i].Condition)
-	}
-
-	return newFilters
-}
-
-// AreFiltersSorted determines whether the expressions in a FiltersExpr are
-// ordered by their expression IDs.
-func (c *CustomFuncs) AreFiltersSorted(f memo.FiltersExpr) bool {
-	for i, n := 0, f.ChildCount(); i < n-1; i++ {
-		if f.Child(i).Child(0).(opt.ScalarExpr).ID() > f.Child(i+1).Child(0).(opt.ScalarExpr).ID() {
-			return false
-		}
-	}
-	return true
-}
-
-// SortFilters sorts a filter list by the IDs of the expressions. This has the
-// effect of canonicalizing FiltersExprs which may have the same filters, but
-// in a different order.
-func (c *CustomFuncs) SortFilters(f memo.FiltersExpr) memo.FiltersExpr {
-	result := make(memo.FiltersExpr, len(f))
-	for i, n := 0, f.ChildCount(); i < n; i++ {
-		fi := f.Child(i).(*memo.FiltersItem)
-		result[i] = *fi
-	}
-	sort.Slice(result, func(i, j int) bool {
-		return result[i].Child(0).(opt.ScalarExpr).ID() < result[j].Child(0).(opt.ScalarExpr).ID()
-	})
-	return result
-}
-
 // ----------------------------------------------------------------------
 //
 // Project functions
-//   General custom match and replace functions used to test and construct
-//   Project and Projections operators.
+//   General functions related to Project operators.
 //
 // ----------------------------------------------------------------------
-
-// CanMergeProjections returns true if the outer Projections operator never
-// references any of the inner Projections columns. If true, then the outer does
-// not depend on the inner, and the two can be merged into a single set.
-func (c *CustomFuncs) CanMergeProjections(outer, inner memo.ProjectionsExpr) bool {
-	innerCols := c.ProjectionCols(inner)
-	for i := range outer {
-		if outer[i].ScalarProps(c.mem).OuterCols.Intersects(innerCols) {
-			return false
-		}
-	}
-	return true
-}
-
-// MergeProjections concatenates the synthesized columns from the outer
-// Projections operator, and the synthesized columns from the inner Projections
-// operator that are passed through by the outer. Note that the outer
-// synthesized columns must never contain references to the inner synthesized
-// columns; this can be verified by first calling CanMergeProjections.
-func (c *CustomFuncs) MergeProjections(
-	outer, inner memo.ProjectionsExpr, passthrough opt.ColSet,
-) memo.ProjectionsExpr {
-	// No need to recompute properties on the new projections, since they should
-	// still be valid.
-	newProjections := make(memo.ProjectionsExpr, len(outer), len(outer)+len(inner))
-	copy(newProjections, outer)
-	for i := range inner {
-		item := &inner[i]
-		if passthrough.Contains(int(item.Col)) {
-			newProjections = append(newProjections, *item)
-		}
-	}
-	return newProjections
-}
-
-// MergeProjectWithValues merges a Project operator with its input Values
-// operator. This is only possible in certain circumstances, which are described
-// in the MergeProjectWithValues rule comment.
-//
-// Values columns that are part of the Project passthrough columns are retained
-// in the final Values operator, and Project synthesized columns are added to
-// it. Any unreferenced Values columns are discarded. For example:
-//
-//   SELECT column1, 3 FROM (VALUES (1, 2))
-//   =>
-//   (VALUES (1, 3))
-//
-func (c *CustomFuncs) MergeProjectWithValues(
-	projections memo.ProjectionsExpr, passthrough opt.ColSet, input memo.RelExpr,
-) memo.RelExpr {
-	newExprs := make(memo.ScalarListExpr, 0, len(projections)+passthrough.Len())
-	newTypes := make([]types.T, 0, len(newExprs))
-	newCols := make(opt.ColList, 0, len(newExprs))
-
-	values := input.(*memo.ValuesExpr)
-	tuple := values.Rows[0].(*memo.TupleExpr)
-	for i, colID := range values.Cols {
-		if passthrough.Contains(int(colID)) {
-			newExprs = append(newExprs, tuple.Elems[i])
-			newTypes = append(newTypes, tuple.Elems[i].DataType())
-			newCols = append(newCols, colID)
-		}
-	}
-
-	for i := range projections {
-		item := &projections[i]
-		newExprs = append(newExprs, item.Element)
-		newTypes = append(newTypes, item.Element.DataType())
-		newCols = append(newCols, item.Col)
-	}
-
-	rows := memo.ScalarListExpr{c.f.ConstructTuple(newExprs, types.TTuple{Types: newTypes})}
-	return c.f.ConstructValues(rows, &memo.ValuesPrivate{
-		Cols: newCols,
-		ID:   values.ID,
-	})
-}
 
 // ProjectionCols returns the ids of the columns synthesized by the given
 // Projections operator.
@@ -769,34 +601,6 @@ func (c *CustomFuncs) ProjectionCols(projections memo.ProjectionsExpr) opt.ColSe
 		colSet.Add(int(projections[i].Col))
 	}
 	return colSet
-}
-
-// ProjectionOuterCols returns the union of all outer columns from the given
-// projection expressions.
-func (c *CustomFuncs) ProjectionOuterCols(projections memo.ProjectionsExpr) opt.ColSet {
-	var colSet opt.ColSet
-	for i := range projections {
-		colSet.UnionWith(projections[i].ScalarProps(c.mem).OuterCols)
-	}
-	return colSet
-}
-
-// AreProjectionsCorrelated returns true if any element in the projections
-// references any of the given columns.
-func (c *CustomFuncs) AreProjectionsCorrelated(
-	projections memo.ProjectionsExpr, cols opt.ColSet,
-) bool {
-	for i := range projections {
-		if projections[i].ScalarProps(c.mem).OuterCols.Intersects(cols) {
-			return true
-		}
-	}
-	return false
-}
-
-// MakeEmptyColSet returns a column set with no columns in it.
-func (c *CustomFuncs) MakeEmptyColSet() opt.ColSet {
-	return opt.ColSet{}
 }
 
 // ProjectExtraCol constructs a new Project operator that passes through all
@@ -814,76 +618,10 @@ func (c *CustomFuncs) ProjectExtraCol(
 
 // ----------------------------------------------------------------------
 //
-// Select Rules
-//   Custom match and replace functions used with select.opt rules.
+// Values functions
+//   General functions related to Values operators.
 //
 // ----------------------------------------------------------------------
-
-// SimplifyFilters removes True operands from a FiltersExpr, and normalizes any
-// False or Null condition to a single False condition. Null values map to False
-// because FiltersExpr are only used by Select and Join, both of which treat a
-// Null filter conjunct exactly as if it were false.
-//
-// SimplifyFilters also "flattens" any And operator child by merging its
-// conditions into a new FiltersExpr list. If, after simplification, no operands
-// remain, then SimplifyFilters returns an empty FiltersExpr.
-//
-// This method assumes that the NormalizeNestedAnds rule has already run and
-// ensured a left deep And tree. If not (maybe because it's a testing scenario),
-// then this rule may rematch, but it should still make forward progress).
-func (c *CustomFuncs) SimplifyFilters(filters memo.FiltersExpr) memo.FiltersExpr {
-	// Start by counting the number of conjuncts that will be flattened so that
-	// the capacity of the FiltersExpr list can be determined.
-	cnt := 0
-	for _, item := range filters {
-		cnt++
-		condition := item.Condition
-		for condition.Op() == opt.AndOp {
-			cnt++
-			condition = condition.(*memo.AndExpr).Left
-		}
-	}
-
-	// Construct new filter list.
-	newFilters := make(memo.FiltersExpr, 0, cnt)
-	for _, item := range filters {
-		var ok bool
-		if newFilters, ok = c.addConjuncts(item.Condition, newFilters); !ok {
-			return memo.FalseFilter
-		}
-	}
-
-	return newFilters
-}
-
-// addConjuncts recursively walks a scalar expression as long as it continues to
-// find nested And operators. It adds any conjuncts (ignoring True operators) to
-// the given FiltersExpr and returns true. If it finds a False or Null operator,
-// it propagates a false return value all the up the call stack, and
-// SimplifyFilters maps that to a FiltersExpr that is always false.
-func (c *CustomFuncs) addConjuncts(
-	scalar opt.ScalarExpr, filters memo.FiltersExpr,
-) (_ memo.FiltersExpr, ok bool) {
-	switch t := scalar.(type) {
-	case *memo.AndExpr:
-		var ok bool
-		if filters, ok = c.addConjuncts(t.Left, filters); !ok {
-			return nil, false
-		}
-		return c.addConjuncts(t.Right, filters)
-
-	case *memo.FalseExpr, *memo.NullExpr:
-		// Filters expression evaluates to False if any operand is False or Null.
-		return nil, false
-
-	case *memo.TrueExpr:
-		// Filters operator skips True operands.
-
-	default:
-		filters = append(filters, memo.FiltersItem{Condition: t})
-	}
-	return filters, true
-}
 
 // ConstructEmptyValues constructs a Values expression with no rows.
 func (c *CustomFuncs) ConstructEmptyValues(cols opt.ColSet) memo.RelExpr {
@@ -899,51 +637,22 @@ func (c *CustomFuncs) ConstructEmptyValues(cols opt.ColSet) memo.RelExpr {
 
 // ----------------------------------------------------------------------
 //
-// GroupBy Rules
-//   Custom match and replace functions used with groupby.opt rules.
+// Grouping functions
+//   General functions related to grouping expressions such as GroupBy,
+//   DistinctOn, etc.
 //
 // ----------------------------------------------------------------------
 
-// AggregationOuterCols returns the union of all outer columns from the given
-// aggregation expressions.
-func (c *CustomFuncs) AggregationOuterCols(aggregations memo.AggregationsExpr) opt.ColSet {
-	var colSet opt.ColSet
-	for i := range aggregations {
-		colSet.UnionWith(aggregations[i].ScalarProps(c.mem).OuterCols)
+// AddColsToGrouping returns a new GroupByDef that is a copy of the given
+// GroupingPrivate, except with the given set of grouping columns union'ed with
+// the existing grouping columns.
+func (c *CustomFuncs) AddColsToGrouping(
+	private *memo.GroupingPrivate, groupingCols opt.ColSet,
+) *memo.GroupingPrivate {
+	return &memo.GroupingPrivate{
+		GroupingCols: private.GroupingCols.Union(groupingCols),
+		Ordering:     private.Ordering,
 	}
-	return colSet
-}
-
-// GroupingAndConstCols returns the grouping columns and ConstAgg columns (for
-// which the input and output column IDs match). A filter on these columns can
-// be pushed through a GroupBy.
-func (c *CustomFuncs) GroupingAndConstCols(
-	grouping *memo.GroupingPrivate, aggs memo.AggregationsExpr,
-) opt.ColSet {
-	result := grouping.GroupingCols.Copy()
-
-	// Add any ConstAgg columns.
-	for i := range aggs {
-		item := &aggs[i]
-		if constAgg, ok := item.Agg.(*memo.ConstAggExpr); ok {
-			// Verify that the input and output column IDs match.
-			if item.Col == constAgg.Input.(*memo.VariableExpr).Col {
-				result.Add(int(item.Col))
-			}
-		}
-	}
-	return result
-}
-
-// GroupingColsAreKey returns true if the input expression's grouping columns
-// form a strict key for its output rows. A strict key means that any two rows
-// will have unique key column values. Nulls are treated as equal to one another
-// (i.e. no duplicate nulls allowed). Having a strict key means that the set of
-// key column values uniquely determine the values of all other columns in the
-// relation.
-func (c *CustomFuncs) GroupingColsAreKey(grouping *memo.GroupingPrivate, input memo.RelExpr) bool {
-	colSet := grouping.GroupingCols
-	return input.Relational().FuncDeps.ColsAreStrictKey(colSet)
 }
 
 // IsUnorderedGrouping returns true if the given grouping ordering is not
@@ -952,461 +661,10 @@ func (c *CustomFuncs) IsUnorderedGrouping(grouping *memo.GroupingPrivate) bool {
 	return grouping.Ordering.Any()
 }
 
-// ----------------------------------------------------------------------
-//
-// Limit Rules
-//   Custom match and replace functions used with limit.opt rules.
-//
-// ----------------------------------------------------------------------
-
-// LimitGeMaxRows returns true if the given constant limit value is greater than
-// or equal to the max number of rows returned by the input expression.
-func (c *CustomFuncs) LimitGeMaxRows(limit tree.Datum, input memo.RelExpr) bool {
-	limitVal := int64(*limit.(*tree.DInt))
-	maxRows := input.Relational().Cardinality.Max
-	return limitVal >= 0 && maxRows < math.MaxUint32 && limitVal >= int64(maxRows)
-}
-
-// ----------------------------------------------------------------------
-//
-// ProjectSet Rules
-//   Custom match and replace functions used with ProjectSet rules.
-//
-// ----------------------------------------------------------------------
-
-// IsZipCorrelated returns true if any element in the zip references
-// any of the given columns.
-func (c *CustomFuncs) IsZipCorrelated(zip memo.ZipExpr, cols opt.ColSet) bool {
-	for i := range zip {
-		if zip[i].ScalarProps(c.mem).OuterCols.Intersects(cols) {
-			return true
-		}
-	}
-	return false
-}
-
-// ZipOuterCols returns the union of all outer columns from the given
-// zip expressions.
-func (c *CustomFuncs) ZipOuterCols(zip memo.ZipExpr) opt.ColSet {
-	var colSet opt.ColSet
-	for i := range zip {
-		colSet.UnionWith(zip[i].ScalarProps(c.mem).OuterCols)
-	}
-	return colSet
-}
-
-// ----------------------------------------------------------------------
-//
-// Set Rules
-//   Custom match and replace functions used with set.opt rules.
-//
-// ----------------------------------------------------------------------
-
-// ProjectColMapLeft returns a Projections operator that maps the left side
-// columns in a SetPrivate to the output columns in it. Useful for replacing set
-// operations with simpler constructs.
-func (c *CustomFuncs) ProjectColMapLeft(set *memo.SetPrivate) memo.ProjectionsExpr {
-	return c.projectColMapSide(set.OutCols, set.LeftCols)
-}
-
-// ProjectColMapRight returns a Project operator that maps the right side
-// columns in a SetPrivate to the output columns in it. Useful for replacing set
-// operations with simpler constructs.
-func (c *CustomFuncs) ProjectColMapRight(set *memo.SetPrivate) memo.ProjectionsExpr {
-	return c.projectColMapSide(set.OutCols, set.RightCols)
-}
-
-// projectColMapSide implements the side-agnostic logic from ProjectColMapLeft
-// and ProjectColMapRight.
-func (c *CustomFuncs) projectColMapSide(toList, fromList opt.ColList) memo.ProjectionsExpr {
-	items := make(memo.ProjectionsExpr, len(toList))
-	for idx, fromCol := range fromList {
-		toCol := toList[idx]
-		items[idx].Element = c.f.ConstructVariable(fromCol)
-		items[idx].Col = toCol
-	}
-	return items
-}
-
-// ----------------------------------------------------------------------
-//
-// Boolean Rules
-//   Custom match and replace functions used with bool.opt rules.
-//
-// ----------------------------------------------------------------------
-
-// ConcatLeftDeepAnds concatenates any left-deep And expressions in the right
-// expression with any left-deep And expressions in the left expression. The
-// result is a combined left-deep And expression. Note that NormalizeNestedAnds
-// has already guaranteed that both inputs will already be left-deep.
-func (c *CustomFuncs) ConcatLeftDeepAnds(left, right opt.ScalarExpr) opt.ScalarExpr {
-	if and, ok := right.(*memo.AndExpr); ok {
-		return c.f.ConstructAnd(c.ConcatLeftDeepAnds(left, and.Left), and.Right)
-	}
-	return c.f.ConstructAnd(left, right)
-}
-
-// NegateComparison negates a comparison op like:
-//   a.x = 5
-// to:
-//   a.x <> 5
-func (c *CustomFuncs) NegateComparison(
-	cmp opt.Operator, left, right opt.ScalarExpr,
-) opt.ScalarExpr {
-	negate := opt.NegateOpMap[cmp]
-	return c.f.DynamicConstruct(negate, left, right).(opt.ScalarExpr)
-}
-
-// CommuteInequality swaps the operands of an inequality comparison expression,
-// changing the operator to compensate:
-//   5 < x
-// to:
-//   x > 5
-func (c *CustomFuncs) CommuteInequality(
-	op opt.Operator, left, right opt.ScalarExpr,
-) opt.ScalarExpr {
-	switch op {
-	case opt.GeOp:
-		return c.f.ConstructLe(right, left)
-	case opt.GtOp:
-		return c.f.ConstructLt(right, left)
-	case opt.LeOp:
-		return c.f.ConstructGe(right, left)
-	case opt.LtOp:
-		return c.f.ConstructGt(right, left)
-	}
-	panic(pgerror.NewAssertionErrorf("called commuteInequality with operator %s", log.Safe(op)))
-}
-
-// FindRedundantConjunct takes the left and right operands of an Or operator as
-// input. It examines each conjunct from the left expression and determines
-// whether it appears as a conjunct in the right expression. If so, it returns
-// the matching conjunct. Otherwise, it returns nil. For example:
-//
-//   A OR A                               =>  A
-//   B OR A                               =>  nil
-//   A OR (A AND B)                       =>  A
-//   (A AND B) OR (A AND C)               =>  A
-//   (A AND B AND C) OR (A AND (D OR E))  =>  A
-//
-// Once a redundant conjunct has been found, it is extracted via a call to the
-// ExtractRedundantConjunct function. Redundant conjuncts are extracted from
-// multiple nested Or operators by repeated application of these functions.
-func (c *CustomFuncs) FindRedundantConjunct(left, right opt.ScalarExpr) opt.ScalarExpr {
-	// Recurse over each conjunct from the left expression and determine whether
-	// it's redundant.
-	for {
-		// Assume a left-deep And expression tree normalized by NormalizeNestedAnds.
-		if and, ok := left.(*memo.AndExpr); ok {
-			if c.isConjunct(and.Right, right) {
-				return and.Right
-			}
-			left = and.Left
-		} else {
-			if c.isConjunct(left, right) {
-				return left
-			}
-			return nil
-		}
-	}
-}
-
-// isConjunct returns true if the candidate expression is a conjunct within the
-// given conjunction. The conjunction is assumed to be left-deep (normalized by
-// the NormalizeNestedAnds rule).
-func (c *CustomFuncs) isConjunct(candidate, conjunction opt.ScalarExpr) bool {
-	for {
-		if and, ok := conjunction.(*memo.AndExpr); ok {
-			if and.Right == candidate {
-				return true
-			}
-			conjunction = and.Left
-		} else {
-			return conjunction == candidate
-		}
-	}
-}
-
-// ExtractRedundantConjunct extracts a redundant conjunct from an Or expression,
-// and returns an And of the conjunct with the remaining Or expression (a
-// logically equivalent expression). For example:
-//
-//   (A AND B) OR (A AND C)  =>  A AND (B OR C)
-//
-// If extracting the conjunct from one of the OR conditions would result in an
-// empty condition, the conjunct itself is returned (a logically equivalent
-// expression). For example:
-//
-//   A OR (A AND B)  =>  A
-//
-// These transformations are useful for finding a conjunct that can be pushed
-// down in the query tree. For example, if the redundant conjunct A is fully
-// bound by one side of a join, it can be pushed through the join, even if B and
-// C cannot.
-func (c *CustomFuncs) ExtractRedundantConjunct(
-	conjunct, left, right opt.ScalarExpr,
-) opt.ScalarExpr {
-	if conjunct == left || conjunct == right {
-		return conjunct
-	}
-
-	return c.f.ConstructAnd(
-		conjunct,
-		c.f.ConstructOr(
-			c.extractConjunct(conjunct, left.(*memo.AndExpr)),
-			c.extractConjunct(conjunct, right.(*memo.AndExpr)),
-		),
-	)
-}
-
-// extractConjunct traverses the And subtree looking for the given conjunct,
-// which must be present. Once it's located, it's removed from the tree, and
-// the remaining expression is returned.
-func (c *CustomFuncs) extractConjunct(conjunct opt.ScalarExpr, and *memo.AndExpr) opt.ScalarExpr {
-	if and.Right == conjunct {
-		return and.Left
-	}
-	if and.Left == conjunct {
-		return and.Right
-	}
-	return c.f.ConstructAnd(c.extractConjunct(conjunct, and.Left.(*memo.AndExpr)), and.Right)
-}
-
-// ----------------------------------------------------------------------
-//
-// Comparison Rules
-//   Custom match and replace functions used with comp.opt rules.
-//
-// ----------------------------------------------------------------------
-
-// NormalizeTupleEquality remaps the elements of two tuples compared for
-// equality, like this:
-//   (a, b, c) = (x, y, z)
-// into this:
-//   (a = x) AND (b = y) AND (c = z)
-func (c *CustomFuncs) NormalizeTupleEquality(left, right memo.ScalarListExpr) opt.ScalarExpr {
-	if len(left) != len(right) {
-		panic(pgerror.NewAssertionErrorf("tuple length mismatch"))
-	}
-	if len(left) == 0 {
-		// () = (), which is always true.
-		return memo.TrueSingleton
-	}
-
-	var result opt.ScalarExpr
-	for i := range left {
-		eq := c.f.ConstructEq(left[i], right[i])
-		if result == nil {
-			result = eq
-		} else {
-			result = c.f.ConstructAnd(result, eq)
-		}
-	}
-	return result
-}
-
-// ----------------------------------------------------------------------
-//
-// Scalar Rules
-//   Custom match and replace functions used with scalar.opt rules.
-//
-// ----------------------------------------------------------------------
-
-// SimplifyCoalesce discards any leading null operands, and then if the next
-// operand is a constant, replaces with that constant.
-func (c *CustomFuncs) SimplifyCoalesce(args memo.ScalarListExpr) opt.ScalarExpr {
-	for i := 0; i < len(args)-1; i++ {
-		item := args[i]
-
-		// If item is not a constant value, then its value may turn out to be
-		// null, so no more folding. Return operands from then on.
-		if !c.IsConstValueOrTuple(item) {
-			return c.f.ConstructCoalesce(args[i:])
-		}
-
-		if item.Op() != opt.NullOp {
-			return item
-		}
-	}
-
-	// All operands up to the last were null (or the last is the only operand),
-	// so return the last operand without the wrapping COALESCE function.
-	return args[len(args)-1]
-}
-
-// AllowNullArgs returns true if the binary operator with the given inputs
-// allows one of those inputs to be null. If not, then the binary operator will
-// simply be replaced by null.
-func (c *CustomFuncs) AllowNullArgs(op opt.Operator, left, right opt.ScalarExpr) bool {
-	return memo.BinaryAllowsNullArgs(op, left.DataType(), right.DataType())
-}
-
-// FoldNullUnary replaces the unary operator with a typed null value having the
-// same type as the unary operator would have.
-func (c *CustomFuncs) FoldNullUnary(op opt.Operator, input opt.ScalarExpr) opt.ScalarExpr {
-	return c.f.ConstructNull(memo.InferUnaryType(op, input.DataType()))
-}
-
-// FoldNullBinary replaces the binary operator with a typed null value having
-// the same type as the binary operator would have.
-func (c *CustomFuncs) FoldNullBinary(op opt.Operator, left, right opt.ScalarExpr) opt.ScalarExpr {
-	return c.f.ConstructNull(memo.InferBinaryType(op, left.DataType(), right.DataType()))
-}
-
-// IsJSONScalar returns if the JSON value is a number, string, true, false, or null.
-func (c *CustomFuncs) IsJSONScalar(value opt.ScalarExpr) bool {
-	v := value.(*memo.ConstExpr).Value.(*tree.DJSON)
-	return v.JSON.Type() != json.ObjectJSONType && v.JSON.Type() != json.ArrayJSONType
-}
-
-// MakeSingleKeyJSONObject returns a JSON object with one entry, mapping key to value.
-func (c *CustomFuncs) MakeSingleKeyJSONObject(key, value opt.ScalarExpr) opt.ScalarExpr {
-	k := key.(*memo.ConstExpr).Value.(*tree.DString)
-	v := value.(*memo.ConstExpr).Value.(*tree.DJSON)
-
-	builder := json.NewObjectBuilder(1)
-	builder.Add(string(*k), v.JSON)
-	j := builder.Build()
-
-	return c.f.ConstructConst(&tree.DJSON{JSON: j})
-}
-
-// IsConstValueEqual returns whether const1 and const2 are equal.
-func (c *CustomFuncs) IsConstValueEqual(const1, const2 opt.ScalarExpr) bool {
-	op1 := const1.Op()
-	op2 := const2.Op()
-	if op1 != op2 || op1 == opt.NullOp {
-		return false
-	}
-	switch op1 {
-	case opt.TrueOp, opt.FalseOp:
-		return true
-	case opt.ConstOp:
-		datum1 := const1.(*memo.ConstExpr).Value
-		datum2 := const2.(*memo.ConstExpr).Value
-		return datum1.Compare(c.f.evalCtx, datum2) == 0
-	default:
-		panic(pgerror.NewAssertionErrorf("unexpected Op type: %v", log.Safe(op1)))
-	}
-}
-
-// SimplifyWhens removes known unreachable WHEN cases and constructs a new CASE
-// statement. Any known true condition is converted to the ELSE. If only the
-// ELSE remains, its expression is returned. condition must be a ConstValue.
-func (c *CustomFuncs) SimplifyWhens(
-	condition opt.ScalarExpr, whens memo.ScalarListExpr, orElse opt.ScalarExpr,
-) opt.ScalarExpr {
-	newWhens := make(memo.ScalarListExpr, 0, len(whens))
-	for _, item := range whens {
-		when := item.(*memo.WhenExpr)
-		if opt.IsConstValueOp(when.Condition) {
-			if !c.IsConstValueEqual(condition, when.Condition) {
-				// Ignore known unmatching conditions.
-				continue
-			}
-
-			// If this is true, we won't ever match anything else, so convert this to
-			// the ELSE (or just return it if there are no earlier items).
-			if len(newWhens) == 0 {
-				return c.ensureTyped(when.Value, memo.InferWhensType(whens, orElse))
-			}
-			return c.f.ConstructCase(condition, newWhens, when.Value)
-		}
-
-		newWhens = append(newWhens, when)
-	}
-
-	// The ELSE value.
-	if len(newWhens) == 0 {
-		// ELSE is the only clause (there are no WHENs), remove the CASE.
-		// NULLs in this position will not be typed, so we tag them with
-		// a type we observed earlier.
-		// typ will never be nil here because the definition of
-		// SimplifyCaseWhenConstValue ensures that whens is nonempty.
-		return c.ensureTyped(orElse, memo.InferWhensType(whens, orElse))
-	}
-
-	return c.f.ConstructCase(condition, newWhens, orElse)
-}
-
-// ensureTyped makes sure that any NULL passing through gets tagged with an
-// appropriate type.
-func (c *CustomFuncs) ensureTyped(d opt.ScalarExpr, typ types.T) opt.ScalarExpr {
-	if d.DataType() == types.Unknown {
-		return c.f.ConstructNull(typ)
-	}
-	return d
-}
-
-// OpsAreSame returns true if the two operators are the same.
-func (c *CustomFuncs) OpsAreSame(left, right opt.Operator) bool {
-	return left == right
-}
-
-// IsConstArray returns true if the expression is a constant array.
-func (c *CustomFuncs) IsConstArray(scalar opt.ScalarExpr) bool {
-	if cnst, ok := scalar.(*memo.ConstExpr); ok {
-		if _, ok := cnst.Value.(*tree.DArray); ok {
-			return true
-		}
-	}
-	return false
-}
-
-// ConvertConstArrayToTuple converts a constant ARRAY datum to the equivalent
-// homogeneous tuple, so ARRAY[1, 2, 3] becomes (1, 2, 3).
-func (c *CustomFuncs) ConvertConstArrayToTuple(scalar opt.ScalarExpr) opt.ScalarExpr {
-	darr := scalar.(*memo.ConstExpr).Value.(*tree.DArray)
-	elems := make(memo.ScalarListExpr, len(darr.Array))
-	ts := make([]types.T, len(darr.Array))
-	for i, delem := range darr.Array {
-		elems[i] = c.f.ConstructConstVal(delem, delem.ResolvedType())
-		ts[i] = darr.ParamTyp
-	}
-	return c.f.ConstructTuple(elems, types.TTuple{Types: ts})
-}
-
-// CastToCollatedString returns the given string or collated string as a
-// collated string constant with the given locale.
-func (c *CustomFuncs) CastToCollatedString(str opt.ScalarExpr, locale string) opt.ScalarExpr {
-	var value string
-	switch t := str.(*memo.ConstExpr).Value.(type) {
-	case *tree.DString:
-		value = string(*t)
-	case *tree.DCollatedString:
-		value = t.Contents
-	default:
-		panic(pgerror.NewAssertionErrorf("unexpected type for COLLATE: %T", log.Safe(str.(*memo.ConstExpr).Value)))
-	}
-
-	d, err := tree.NewDCollatedString(value, locale, &c.f.evalCtx.CollationEnv)
-	if err != nil {
-		panic(err)
-	}
-	return c.f.ConstructConst(d)
-}
-
-// MakeUnorderedSubquery returns a SubqueryPrivate that specifies no ordering.
-func (c *CustomFuncs) MakeUnorderedSubquery() *memo.SubqueryPrivate {
-	return &memo.SubqueryPrivate{}
-}
-
-// SubqueryOrdering returns the ordering property on a SubqueryPrivate.
-func (c *CustomFuncs) SubqueryOrdering(sub *memo.SubqueryPrivate) physical.OrderingChoice {
-	var oc physical.OrderingChoice
-	oc.FromOrdering(sub.Ordering)
-	return oc
-}
-
-// FirstCol returns the first column in the input expression.
-func (c *CustomFuncs) FirstCol(in memo.RelExpr) opt.ColumnID {
-	inCol, _ := c.OutputCols(in).Next(0)
-	return opt.ColumnID(inCol)
-}
-
-// MakeArrayAggCol returns a ColPrivate with the given type and an "array_agg" label.
-func (c *CustomFuncs) MakeArrayAggCol(typ types.T) *memo.ColPrivate {
-	return &memo.ColPrivate{Col: c.mem.Metadata().AddColumn("array_agg", typ)}
+// MakeGrouping constructs a new unordered GroupingPrivate using the given
+// grouping columns.
+func (c *CustomFuncs) MakeGrouping(groupingCols opt.ColSet) *memo.GroupingPrivate {
+	return &memo.GroupingPrivate{GroupingCols: groupingCols}
 }
 
 // MakeOrderedGrouping constructs a new GroupingPrivate using the given
@@ -1417,35 +675,25 @@ func (c *CustomFuncs) MakeOrderedGrouping(
 	return &memo.GroupingPrivate{GroupingCols: groupingCols, Ordering: ordering}
 }
 
-// IsLimited indicates whether a limit was pushed under the subquery
-// already. See e.g. the rule IntroduceExistsLimit.
-func (c *CustomFuncs) IsLimited(sub *memo.SubqueryPrivate) bool {
-	return sub.WasLimited
-}
-
-// MakeLimited specifies that the subquery has a limit set
-// already. This prevents e.g. the rule IntroduceExistsLimit from
-// applying twice.
-func (c *CustomFuncs) MakeLimited(sub *memo.SubqueryPrivate) *memo.SubqueryPrivate {
-	newSub := *sub
-	newSub.WasLimited = true
-	return &newSub
+// ExtractGroupingOrdering returns the ordering associated with the input
+// GroupingPrivate.
+func (c *CustomFuncs) ExtractGroupingOrdering(
+	private *memo.GroupingPrivate,
+) physical.OrderingChoice {
+	return private.Ordering
 }
 
 // ----------------------------------------------------------------------
 //
-// Numeric Rules
-//   Custom match and replace functions used with numeric.opt rules.
+// Constant value functions
+//   General functions related to constant values and datums.
 //
 // ----------------------------------------------------------------------
 
-// IsAdditive returns true if the type of the expression supports addition and
-// subtraction in the natural way. This differs from "has a +/- Numeric
-// implementation" because JSON has an implementation for "- INT" which doesn't
-// obey x - 0 = x. Additive types include all numeric types as well as
-// timestamps and dates.
-func (c *CustomFuncs) IsAdditive(e opt.ScalarExpr) bool {
-	return types.IsAdditiveType(e.DataType())
+// IsPositiveLimit is true if the given limit value is greater than zero.
+func (c *CustomFuncs) IsPositiveLimit(limit tree.Datum) bool {
+	limitVal := int64(*limit.(*tree.DInt))
+	return limitVal > 0
 }
 
 // EqualsNumber returns true if the given numeric value (decimal, float, or
@@ -1469,397 +717,4 @@ func (c *CustomFuncs) EqualsNumber(datum tree.Datum, value int64) bool {
 		return *t == tree.DInt(value)
 	}
 	return false
-}
-
-// ----------------------------------------------------------------------
-//
-// Constant Folding Rules
-//   Custom match and replace functions used with fold_constants.opt
-//   rules.
-//
-// ----------------------------------------------------------------------
-
-// IsListOfConstants returns true if elems is a list of constant values or
-// tuples.
-func (c *CustomFuncs) IsListOfConstants(elems memo.ScalarListExpr) bool {
-	for _, elem := range elems {
-		if !c.IsConstValueOrTuple(elem) {
-			return false
-		}
-	}
-	return true
-}
-
-// FoldArray evaluates an Array expression with constant inputs. It returns the
-// array as a Const datum with type TArray.
-func (c *CustomFuncs) FoldArray(elems memo.ScalarListExpr, typ types.T) opt.ScalarExpr {
-	elemType := typ.(types.TArray).Typ
-	a := tree.NewDArray(elemType)
-	a.Array = make(tree.Datums, len(elems))
-	for i := range a.Array {
-		a.Array[i] = memo.ExtractConstDatum(elems[i])
-		if a.Array[i] == tree.DNull {
-			a.HasNulls = true
-		}
-	}
-	return c.f.ConstructConst(a)
-}
-
-// IsConstValueOrTuple returns true if the input is a constant or a tuple of
-// constants.
-func (c *CustomFuncs) IsConstValueOrTuple(input opt.ScalarExpr) bool {
-	return memo.CanExtractConstDatum(input)
-}
-
-// FoldBinary evaluates a binary expression with constant inputs. It returns
-// a constant expression as long as it finds an appropriate overload function
-// for the given operator and input types, and the evaluation causes no error.
-func (c *CustomFuncs) FoldBinary(op opt.Operator, left, right opt.ScalarExpr) opt.ScalarExpr {
-	lDatum, rDatum := memo.ExtractConstDatum(left), memo.ExtractConstDatum(right)
-
-	o, ok := memo.FindBinaryOverload(op, left.DataType(), right.DataType())
-	if !ok {
-		return nil
-	}
-
-	result, err := o.Fn(c.f.evalCtx, lDatum, rDatum)
-	if err != nil {
-		return nil
-	}
-	return c.f.ConstructConstVal(result, o.ReturnType)
-}
-
-// FoldUnary evaluates a unary expression with a constant input. It returns
-// a constant expression as long as it finds an appropriate overload function
-// for the given operator and input type, and the evaluation causes no error.
-func (c *CustomFuncs) FoldUnary(op opt.Operator, input opt.ScalarExpr) opt.ScalarExpr {
-	datum := memo.ExtractConstDatum(input)
-
-	o, ok := memo.FindUnaryOverload(op, input.DataType())
-	if !ok {
-		return nil
-	}
-
-	result, err := o.Fn(c.f.evalCtx, datum)
-	if err != nil {
-		return nil
-	}
-	return c.f.ConstructConstVal(result, o.ReturnType)
-}
-
-// FoldCast evaluates a cast expression with a constant input. It returns
-// a constant expression as long as the evaluation causes no error.
-func (c *CustomFuncs) FoldCast(input opt.ScalarExpr, colType coltypes.T) opt.ScalarExpr {
-	switch colType.(type) {
-	case *coltypes.TOid:
-		// Save this cast for the execbuilder.
-		return nil
-	}
-
-	datum := memo.ExtractConstDatum(input)
-	texpr, err := tree.NewTypedCastExpr(datum, colType)
-	if err != nil {
-		return nil
-	}
-
-	result, err := texpr.Eval(c.f.evalCtx)
-	if err != nil {
-		return nil
-	}
-
-	return c.f.ConstructConstVal(result, c.ColTypeToDatumType(colType))
-}
-
-// isMonotonicConversion returns true if conversion of a value from FROM to
-// TO is monotonic.
-// That is, if a and b are values of type FROM, then
-//
-//   1. a = b implies a::TO = b::TO and
-//   2. a < b implies a::TO <= b::TO
-//
-// Property (1) can be violated by cases like:
-//
-//   '-0'::FLOAT = '0'::FLOAT, but '-0'::FLOAT::STRING != '0'::FLOAT::STRING
-//
-// Property (2) can be violated by cases like:
-//
-//   2 < 10, but  2::STRING > 10::STRING.
-//
-// Note that the stronger version of (2),
-//
-//   a < b implies a::TO < b::TO
-//
-// is not required, for instance this is not generally true of conversion from
-// a TIMESTAMP to a DATE, but certain such conversions can still generate spans
-// in some cases where values under FROM and TO are "the same" (such as where a
-// TIMESTAMP precisely falls on a date boundary).  We don't need this property
-// because we will subsequently check that the values can round-trip to ensure
-// that we don't lose any information by doing the conversion.
-// TODO(justin): fill this out with the complete set of such conversions.
-func isMonotonicConversion(from, to coltypes.T) bool {
-	if from == coltypes.Timestamp ||
-		from == coltypes.TimestampWithTZ ||
-		from == coltypes.Date {
-		return to == coltypes.Timestamp ||
-			to == coltypes.TimestampWithTZ ||
-			to == coltypes.Date
-	}
-
-	if from == coltypes.Int8 ||
-		from == coltypes.Float8 ||
-		from == coltypes.Decimal {
-		return to == coltypes.Int8 ||
-			to == coltypes.Float8 ||
-			to == coltypes.Decimal
-	}
-
-	return false
-}
-
-// UnifyComparison attempts to convert a constant expression to the type of the
-// variable expression, if that conversion can round-trip and is monotonic.
-func (c *CustomFuncs) UnifyComparison(left, right opt.ScalarExpr) opt.ScalarExpr {
-	v := left.(*memo.VariableExpr)
-	cnst := right.(*memo.ConstExpr)
-
-	desiredType := v.DataType()
-	originalType := cnst.DataType()
-
-	// Don't bother if they're already the same.
-	if desiredType.Equivalent(originalType) {
-		return nil
-	}
-
-	desiredColType, err := coltypes.DatumTypeToColumnType(desiredType)
-	if err != nil {
-		return nil
-	}
-
-	originalColType, err := coltypes.DatumTypeToColumnType(originalType)
-	if err != nil {
-		return nil
-	}
-
-	if !isMonotonicConversion(originalColType, desiredColType) {
-		return nil
-	}
-
-	// Check that the datum can round-trip between the types. If this is true, it
-	// means we don't lose any information needed to generate spans, and combined
-	// with monotonicity means that it's safe to convert the RHS to the type of
-	// the LHS.
-	convertedDatum, err := tree.PerformCast(c.f.evalCtx, cnst.Value, desiredColType)
-	if err != nil {
-		return nil
-	}
-
-	convertedBack, err := tree.PerformCast(c.f.evalCtx, convertedDatum, originalColType)
-	if err != nil {
-		return nil
-	}
-
-	if convertedBack.Compare(c.f.evalCtx, cnst.Value) != 0 {
-		return nil
-	}
-
-	return c.f.ConstructConst(convertedDatum)
-}
-
-// FoldComparison evaluates a comparison expression with constant inputs. It
-// returns a constant expression as long as it finds an appropriate overload
-// function for the given operator and input types, and the evaluation causes
-// no error.
-func (c *CustomFuncs) FoldComparison(op opt.Operator, left, right opt.ScalarExpr) opt.ScalarExpr {
-	lDatum, rDatum := memo.ExtractConstDatum(left), memo.ExtractConstDatum(right)
-
-	var flipped, not bool
-	o, flipped, not, ok := memo.FindComparisonOverload(op, left.DataType(), right.DataType())
-	if !ok {
-		return nil
-	}
-
-	if flipped {
-		lDatum, rDatum = rDatum, lDatum
-	}
-
-	result, err := o.Fn(c.f.evalCtx, lDatum, rDatum)
-	if err != nil {
-		return nil
-	}
-	if b, ok := result.(*tree.DBool); ok && not {
-		result = tree.MakeDBool(!*b)
-	}
-	return c.f.ConstructConstVal(result, types.Bool)
-}
-
-// FoldFunction evaluates a function expression with constant inputs. It
-// returns a constant expression as long as the function is contained in the
-// FoldFunctionWhitelist, and the evaluation causes no error.
-func (c *CustomFuncs) FoldFunction(
-	args memo.ScalarListExpr, private *memo.FunctionPrivate,
-) opt.ScalarExpr {
-	if _, ok := FoldFunctionWhitelist[private.Name]; !ok {
-		return nil
-	}
-
-	exprs := make(tree.TypedExprs, len(args))
-	for i := range exprs {
-		exprs[i] = memo.ExtractConstDatum(args[i])
-	}
-	funcRef := tree.WrapFunction(private.Name)
-	fn := tree.NewTypedFuncExpr(
-		funcRef,
-		0, /* aggQualifier */
-		exprs,
-		nil, /* filter */
-		nil, /* windowDef */
-		private.Typ,
-		private.Properties,
-		private.Overload,
-	)
-
-	result, err := fn.Eval(c.f.evalCtx)
-	if err != nil {
-		return nil
-	}
-	return c.f.ConstructConstVal(result, private.Typ)
-}
-
-// FoldFunctionWhitelist contains functions that are known to always produce
-// the same result given the same set of arguments. This excludes impure
-// functions and functions that rely on context such as locale or current user.
-// See sql/sem/builtins/builtins.go for the function definitions.
-// TODO(rytaft): This is a stopgap until #26582 is completed to identify
-// functions as immutable, stable or volatile.
-var FoldFunctionWhitelist = map[string]struct{}{
-	"length":                        {},
-	"char_length":                   {},
-	"character_length":              {},
-	"bit_length":                    {},
-	"octet_length":                  {},
-	"lower":                         {},
-	"upper":                         {},
-	"substr":                        {},
-	"substring":                     {},
-	"concat":                        {},
-	"concat_ws":                     {},
-	"convert_from":                  {},
-	"convert_to":                    {},
-	"to_uuid":                       {},
-	"from_uuid":                     {},
-	"abbrev":                        {},
-	"broadcast":                     {},
-	"family":                        {},
-	"host":                          {},
-	"hostmask":                      {},
-	"masklen":                       {},
-	"netmask":                       {},
-	"set_masklen":                   {},
-	"text":                          {},
-	"inet_same_family":              {},
-	"inet_contained_by_or_equals":   {},
-	"inet_contains_or_contained_by": {},
-	"inet_contains_or_equals":       {},
-	"from_ip":                       {},
-	"to_ip":                         {},
-	"split_part":                    {},
-	"repeat":                        {},
-	"encode":                        {},
-	"decode":                        {},
-	"ascii":                         {},
-	"chr":                           {},
-	"md5":                           {},
-	"sha1":                          {},
-	"sha256":                        {},
-	"sha512":                        {},
-	"fnv32":                         {},
-	"fnv32a":                        {},
-	"fnv64":                         {},
-	"fnv64a":                        {},
-	"crc32ieee":                     {},
-	"crc32c":                        {},
-	"to_hex":                        {},
-	"to_english":                    {},
-	"strpos":                        {},
-	"overlay":                       {},
-	"lpad":                          {},
-	"rpad":                          {},
-	"btrim":                         {},
-	"ltrim":                         {},
-	"rtrim":                         {},
-	"reverse":                       {},
-	"replace":                       {},
-	"translate":                     {},
-	"regexp_extract":                {},
-	"regexp_replace":                {},
-	"like_escape":                   {},
-	"not_like_escape":               {},
-	"ilike_escape":                  {},
-	"not_ilike_escape":              {},
-	"similar_to_escape":             {},
-	"not_similar_to_escape":         {},
-	"initcap":                       {},
-	"quote_ident":                   {},
-	"left":                          {},
-	"right":                         {},
-	"greatest":                      {},
-	"least":                         {},
-	"abs":                           {},
-	"acos":                          {},
-	"asin":                          {},
-	"atan":                          {},
-	"atan2":                         {},
-	"cbrt":                          {},
-	"ceil":                          {},
-	"ceiling":                       {},
-	"cos":                           {},
-	"cot":                           {},
-	"degrees":                       {},
-	"div":                           {},
-	"exp":                           {},
-	"floor":                         {},
-	"isnan":                         {},
-	"ln":                            {},
-	"log":                           {},
-	"mod":                           {},
-	"pi":                            {},
-	"pow":                           {},
-	"power":                         {},
-	"radians":                       {},
-	"round":                         {},
-	"sin":                           {},
-	"sign":                          {},
-	"sqrt":                          {},
-	"tan":                           {},
-	"trunc":                         {},
-	"string_to_array":               {},
-	"array_to_string":               {},
-	"array_length":                  {},
-	"array_lower":                   {},
-	"array_upper":                   {},
-	"array_append":                  {},
-	"array_prepend":                 {},
-	"array_cat":                     {},
-	"json_remove_path":              {},
-	"json_extract_path":             {},
-	"jsonb_extract_path":            {},
-	"json_set":                      {},
-	"jsonb_set":                     {},
-	"jsonb_insert":                  {},
-	"jsonb_pretty":                  {},
-	"json_typeof":                   {},
-	"jsonb_typeof":                  {},
-	"array_to_json":                 {},
-	"to_json":                       {},
-	"to_jsonb":                      {},
-	"json_build_array":              {},
-	"jsonb_build_array":             {},
-	"json_build_object":             {},
-	"jsonb_build_object":            {},
-	"json_object":                   {},
-	"jsonb_object":                  {},
-	"json_strip_nulls":              {},
-	"jsonb_strip_nulls":             {},
-	"json_array_length":             {},
-	"jsonb_array_length":            {},
 }
